Housekeeping: fix re-renders when character URLs are added to context
We add character URLs to the context using an effect when the movie
detail page is loaded. This approach caused unnecessary re-renders
because one of the effect's dependencies was an array (because the
`useEffect` hook uses the `Object.is` algorithm under the hood --which
works by checking object/array references-- the aforementioned effect
would fire on every render).

This commit fixes the issue by memoizing the array dependency before
adding the URLs to the `CharacterUrlContext`.

Additional changes:
* Refactor: rename character url context to better reflect its role